voice_poc/pyproject.toml
[project]
name = "voice-poc"
version = "0.1.0"
description = "Terminal-only Hindi voice agent PoC (ASR -> LLM -> TTS)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.52.0",
    "pydub>=0.25.1",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]

[tool.uv]
index-url = "https://pypi.org/simple"

